rename castlectl to castle
