+++ cmd/castle/node-list_test.go
@@ -1,4 +1,4 @@
-package castlectl
+package castle
 
 import (
 	"fmt"
@@ -7,7 +7,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/quantum/castle/pkg/castlectl/test"
+	"github.com/quantum/castle/pkg/castle/test"
 	"github.com/quantum/castle/pkg/model"
 )
