+++ cmd/castle/node.go
@@ -1,4 +1,4 @@
-package castlectl
+package castle
 
 import "github.com/spf13/cobra"
