+++ cmd/castle/pool-list.go
@@ -1,11 +1,11 @@
-package castlectl
+package castle
 
 import (
 	"bytes"
 	"fmt"
 	"net/http"
 
-	"github.com/quantum/castle/pkg/castlectl/client"
+	"github.com/quantum/castle/pkg/castle/client"
 	"github.com/quantum/castle/pkg/model"
 	"github.com/quantum/castle/pkg/util/display"
 	"github.com/quantum/castle/pkg/util/flags"
